[wolsen,r=] Change contrib.openstack.ip endpoint functions to specify an override of an address rather than the endpoint value itself.
This change makes it easier to use since users will only need to specify
a single public-addres, internal-address, or admin-address on a charm and
it can apply to all endpoints created by a charm.

This method is safe to alter the actual behavior of the method rather than
honoring the legacy behavior since it is not used by any of the openstack
charms to date as the one merge including it has been withdrawn in favor
of using a simpler approach.