unison: Add all known keys for each host
Adding only the rsa key is problematic.  The remote host could be
using ecdsa and the first attmempt to connect will add that ecdsa key
and complain that the rsa key conflicts.  Taking all key types adds
flexibility and future-proofing.

Without this, connecting to a host that presents an ecdsa host key will
warn that the rsa key conflicts and will require interactive input to
accept the ecdsa key; this breaks the charm.