[billy-olsen,r=] Improve Ceph pg num calculations.
This change modifies the calculation in determining the number of
placement groups by following the detailed formula provided by the
ceph PG Calc website. Previously, the calculation was based on the
Ceph documentation for pool sizes, which is misleading as the data
provided in the documentation is for the maximum number of placement
groups for the cluster itself.

This change also adds in an option to specify the percent of data that
is intended to be used by the pool being created. If no information is
provided on the percentage of data for the pool, then this code will
use a default of 10% - a value which may result in low number of
placement groups for the pool but will also result in a setting which
can later be changed by the user (since pgs can be increased but never
decreased).

Partially Closes-Bug: #1492742