Pass an IP to ssh_ip instead of listen_to
Current 'listen_to' setting in hardening is not very useful in cloud environments.
Each unit has its own IP and therefore it's impossible to configure the application with
a single config file. This patch extends functionality and figures out an IP on every
unit and passes it as ssh_ip variable in the template.