[billy-olsen,r=] Allow passing of parameters to upstart scripts.
Some upstart scripts require that additional key/value parameters
are specified in order to narrow down the instance of the service
which should be started/stopped/restarted/etc. This change adds
kwargs parameters to each of the service management functions in
order provide allow any attribute required by the underlying upstart
script to be specified.

A notable limitation is that the service_pause and service_resume
will now allow the passing of parameters to the underlying service
for stopping and starting the service instance. However, all of the
service instances will be manually disabled for non-systemd init
systems.