expect_ha function checks for VIP or dns-ha settings
In OpenStack charms that relate to hacluster it may take time for
hacluster to complete. In the meantime the charm needs a way to be
informed to wait for hacluster. Simply checking for the relation is
too racey. This function checks for tell tale configuration
paramaters. The charm can then react accordingly.

Partial-bug: #1661392