mysql: Ensure sane defaults, fixup tests
The default behaviour for no user provided buffer size configuration is
to use 50% of system memory; in a container, this might be 50% of the
hosting system, which does not make a huge amount of sense.

Default the minimum of 512MB or 50% of ram unless otherwise specified
by the charm user.

Also fixup tests to run under py2 and py3.