Rework default worker-multiplier behaviour
The typical OpenStack deployment places control plane services in LXD
containers, which currently see all of the resources available on the
hosting server.

Current behaviour means that the services are configured with 2x the
detected CPU core count, resulting in high memory usage and contention
when services are deployed on servers with high cpu counts.

Tweak the default behaviour to cap at 4 workers *unless* a value is
actually provided via the worker-multiplier option that each charm has.

Note that this behaviour is only applied when running within a container.