Ensure the config_flags_parser returns an OrderedDict.
The result from the config_flags_parser is often fed into a template
and written in the order of iteration of the returned dict. This causes
problems due to iteration of dicts not being predictable or reliable.
Different iteration ordering may cause inadvertent restarts of services
which are guarded by the restart_on_change decorator.

This patch ensures that config-flags which are in yaml format are returned
in an OrderedDict which is sorted by the keys. Nested dictionaries are
also sorted in order to guarantee the iteration of the entire structure.

Config flags which are not in yaml form, will be returned in the order
that they are parsed. This is safe since the parsing is linear along the
string.

Partial-Bug: #1698343