Remove E402 ingore for pep8 to improve code checking
The flake8 pep8 tests used an --ignore E402 which is about
not having all the imports before 'code'.  This is a known
code QA issue, and so this patch removes the ignore.  One
test module needed to be patched with a #noqa, but that's
because the files imported adjusted their behaviour based
on the changed line for the tests.