Update NeutronAPIContext to handle enable-qos
Update the context used to examine the relationship with the neutron-api
charm so that it reacts to enable-qos being set.

If enable-qos is set then the 'enable_qos' key is added to the context
as well as the updating the extensions setting. Charms are likely to
only be interested in the extensions setting but enable_qos is set for
completeness.