Modulo Distribution (#8)
Modulo Distribution

Charms have needed a way to distribute load during expensive operations
in large scale deployments. This change adds two helpers to mitigate
load.

modulo_distribution uses unit number, a modulo value and a constant wait time
to produce a calculated wait time distribution.

If you have 1000 nodes that need to restart 100 at a time 1 minute at a time:

      time.wait(modulo_distribution(modulo=100, wait=60))
      restart()

If you need restarts to happen serially set modulo to the exact number of
nodes and set a high constant wait time:

      time.wait(modulo_distribution(modulo=10, wait=120))
      restart()

distributed_wait is a helper for OpenStack charms to use which checks
for config values to pass to modulo_distribution and uses the returned
value to use with time.wait().