Fix mismatch between 'internal' and 'int' in DNSHA
DNSHA keys off the string in configuration variables os-*-hostname.
os-internal-hostname is the name of the config variable. However,
resolve_address uses a long standing constant ADDRESS_MAP that refers
to this as 'int'.

This change handles this and hands resolve_address what it expects.
Also, use a clearer variable name for endpoint_type.