Ingress address function (#35)
Ingress address function and  iter_units_for_relation_name()

The new features in Juju set ingress-address and egress-networks on all
relations based on network spaces. Ingress-address should be used
over private-address when it is available. As this is a new feature we
fall back to private-address when ingress-address is not available.

This new function consumes the relation data and returns the
ingress-address when available or the private-address when not.

Add the iter_units_for_relation_name generator for iterating units
based on relation name.