Py3 save script rc (#44)
* Update cinder helpers for v2+ compatability

Cinder volume and snapshot objects use 'name' attribute rather than
'display_name'. This change updates the helpers to deal with this
when creating objects and when retrieving the name attribute.

* Write save_script_rc as text

save_script_rc is failing under py3 ("'str' does not support the
buffer interface") so write the file as text since its a text file.