Add OpenStack Context for identity-credentials interface type
Add new IdentityCredentialsContext based on IdentityServiceContext
which is __call__() return dict compatible with the existing
templates as used with the identity-service relation.

This will be used by the ceilometer charm for OpenStack Queens
or later where access to other services is required, but the
charm won't actually register any endpoints in the keystone
catalog.