add extra-policy context and section
* Extra policy context is added to render service-specific policy
drop-in files (newer yaml or json format);
* section-extra-policy is trivial but is included in service-specific
policy-drop-in templates - this is mainly for naming policy files
differently to allow primary and subordinate charm policy files to
co-exist in a single directory without collisions;
* every classic charm to use this context will need to get extra-policy
config option;
* a policy directory is created if not present (policy.d is used by
default but it is intentionally passed as an __init__ argument to allow
potential overrides.

See pad.lv/1741723.