Add string template rendering to core/templating (#102)
This change allows a charm writer to support specifying
config template parts in config options, or render config files based on
templates passed via relation data from other charms.

The main use-case is drop-in files which contain custom logic not
possible to predict or model in the charm code by providing relevant
config options.

For example, RBAC rules are subjective and some operators have
to implement policy based on those subjective rules which may
differ from defaults. If those rules cannot be provided via an API
to a given service and are stored in config files this is the only way
to provide desired customization options.

Related to pad.lv/1741723
In addition to #87 but for
reactive charms which mainly use this render function.