Add VolumeAPIContext for catalog and version info
Adds a VolumeAPIContext object which provides information
about the volume api that should be in use. The VolumeAPIContext
provides information rearding the catalog_info that can be passed
directly to a configuration file which accepts a catalog_info
string.

Additionally, the volume api version string is returned for any
instances which do not accept the catalog_info string.