Add Volume API Context (#65) (#111)
* Revert "Use volumev3 for Openstack >= Pike (#65) (#66)"

This reverts commit 3aa436da7151770e104df7d2ff013fbca3cb4367.

After discussion with @ajkavanaugh (tinwood) this change conflates
the internal endpoints with api version usage for cinder API. A
new change will be used instead.

* Add VolumeAPIContext for catalog and version info

Adds a VolumeAPIContext object which provides information
about the volume api that should be in use. The VolumeAPIContext
provides information rearding the catalog_info that can be passed
directly to a configuration file which accepts a catalog_info
string.

Additionally, the volume api version string is returned for any
instances which do not accept the catalog_info string.