Read in ca certificate as binary for PY3
The comparison of bytes vs string of the CA certificate produces a
false negative. This leads to rewriting certificates and affecting
connectivity to services. The writing of the file was changed to bytes
in https://github.com/juju/charm-helpers/pull/58/.

Read in the certificate as bytes as well for a bytes vs bytes
comparison.

Closes-Issue: https://github.com/juju/charm-helpers/issues/86
Closes-Bug: #1762431