Cache config-get data
The config() function is currently decorated with the @cached
decorator, however, in a typical charm this will not have
much effect as the charm will make every call to config()
with different arguments, subsequently bypassing the cache.

This commit caches the output from config-get --all and
returns the Config object or each individual value the
caller might ask for.