Bug/1761305/ensure apache ssl (#151)
* Make the detection of which apt pkgs are installed more robust

Basically, the original test for installation didn't account for the
fact that a package might be have been installed and then removed.
This test checks for whether it is actually installed and if not
then installs it.

* Change the perms on Apache SSL files to 640

This is due to related bug in that the SSL files were world readable
and thus, a security risk.  They have been changed to 640 to allow
the user and group to read them.

Related-Bug: #1761305 on launchpad

* Fix side-effect during unit testing of creating ':memory:' file

As an artifact of a recent commit, duing unit testing a file
called ':memory:' was being created.  This commit just checks to see
if the path passed to Storage.__init__() is ':memory:' and doesn't
create the file.