Don't error/raise an exception on port already open
In some cases, we have multiple units deployed to the same machine. An
example of this is nrpe subordinates related to different services
deployed to the same machine. In these special situations, open-port
fails per LP#1750490 and as follows:

| ERROR cannot open 5666/tcp (unit "nrpe-ceph/0"): conflicts with existing 5666/tcp (unit "nrpe-ceph-mon/0")

This fix ignores that error and allows hooks to continue to run.