Allow forcing keystone preferred-api-version
The get_default_keystone_session function is useful for handling the
fact that queens deployments may have preferred-api-version unset.
However, it assumes v2 for < queens.

This change allows the caller to force the preferred api version.

This is useful for the keystone-ldap charm which uses v3 regardless of
OpenStack release.