ovs: long interface names and existing wiring
Linux interface names are a maximum of 15 characters; ensure that
when wiring an ovs bridge to a linux bridge, the generated veth
pair naming is below this limit by using parts of the hash of
the bridge name.

Any existing use of this feature will remain unchanged as this
is only used when the provided bridge name is > 10 characters.

This commit also fixes an issue when a linuxbridge was previously
directly wired into ovs bridge - this was supported in older
code versions, and we need to ensure we don't break users who
have leveraged this feature on upgrades.