Support multiple WSGI vhosts in Openstack (#201)
The current WSGIWorkerConfigContext cannot be used to support
multiple wsgi vhosts for the same user because the WSGIDaemonProcess
'name' attribute must be different between the two vhosts but the
'name' is also used to dictate the user and group. This change
allows the user and group to be explicitly set. The allows for
multiple wsgi vhosts for user X with different WSGIDaemonProcess
names.