Allow glance image hypervisor type to be unset
Explicitly setting a hypervisor image type is mostly useful when
a cloud contains mixed hypervisors. This is not the case in the
existing amulet tests so allow the hypervisor attribute to be unset.

In addition allow a request for a glance client to force a glance
v1 client, this is to work around Bug #1371559