Make the harden and pausable_restart_on_change lazy
Due to py3 work in some charms, these decorators now take callables
in their parameter list as an option, so that load time decorator usage
doesn't have to force a long function call; it makes the parameters lazy
and only evaluated when the decorated function is called.

Also fix a load of linting errors that had been building up.