Extra config when generating Openstack HA settings
When generating the config to send to the hacluster charm some
charms need to add additional services etc to be managed by
hacluster. generate_ha_relation_data passes back settings so a
charm could update them once generated but they are passed back
JSON encoded so rather then having to decode and re-encode this
change allows the charm to pass a dict of settings to be added
to the config prior to encoding.