Update copy_nrpe_checks() for optional c-h directory (#247)
The function assumed that the charmhelpers directory would always be in
the hooks directory.  However, with PY3 it can be very tricky to try to
keep a symlink in the hooks directory.  Therefore, for some PY3
non-reactive charms, charmhelpers was moved to the CHARMDIR.  However,
this caused this function to fail.

The change here is to search for the charmhelpers directory in both
CHARMDIR and CHARMDIR/hooks and then use that.

Closes-Bug: #246
LP-Bug: #1796830