Refactor vip resource creation for iface'less use
Pacemaker VIP resources are able to automatically detect and
configure correct iface and netmask parameters based on local
configuration of the unit.

Switch to using hash based naming in VIP resources and
drop configuration of nic and cidr_netmask params.  The hash
is based on the first 7 chars of an sha1 hash of the VIP
which is deterministic and still readable (base of gits short
hash default).

This supports deployment in container based environments
where interface -> subnet connectivity may not be consistent
between container instances within an application.

NOTE: this change is somewhat bumpy as the original iface named
VIP resource will be stopped and deleted prior to the creation
of the new short hash named VIP resource.

NOTE: This change was mainly authored by jamespage
(https://review.openstack.org/#/c/531359 ) and is ported to
charmhelpers so it can be used across multiple charms.