Improve logging of filesystem permissions
Permissions should always be logged in octal, which wasn't previously
the case for the "Changing permissions on existing content" message.  I
also tweaked the format of a couple of existing messages to add an
explicit "0o" prefix so that nobody needs to look at the code to work
out which base is in use.