Add support for ``app_name`` in ``add_op_create_pool``
Some time ago upstream Ceph releases gained support for tagging
pools with which applications they are destined for.

Use cases are emerging for this and the first real user of it will
be the ``rbd-mirror`` feature enablement work that is ongoing in
the OpenStack Ceph charms.

By adding the ``rbd`` tag to the pools you use with Ceph RBD you
will get a "it just works" experience when adding your second
remote Ceph cluster and relating them using the ``ceph-rbd-mirror``
charm.