Re-enable pgrep_full (#309)
pidof -x is not finding scripts run under python3 (like neutron-server).
pgrep -f has not problem finding them.

We had most of the plumbing already in place to use pgrep. It turns out
pidof is susceptible to the same problems pgrep does which led us to
change in the first place.

This change will fix a number of stein testing enrolment blockers.