Add Contexts for Nova Vendor Metadata
Added two new generic contexts to avoid more duplication
of code on several projects when dealing with nova
vendor metadata.

NovaVendorMetadataContext is responsible for reading
"vendor-data" and "vendor-data-url" config options and
returning the proper context for setting the values to
nova.conf file.

NovaVendorMetadataJSONContext is responsible for reading
only the vendor-data config option, validating it and
returning the context to write the value to vendor_data.json
file.

If a particular charm needs to complement
the logic, such as disregarding a certain value for a
given OpenStack release or propagating the values to a
relation, it is recommended that the context is inherited
and specialized for that particular charm.

Related-Bug: #1777714