Set unit_name when requesting certificates.
A fix has landed in the tls-certificates interface to support
relating to vaults certificates relation over a cross-model
relation *1. The old style charms do not use this interface
so the fixes need to be ported to charm helpers. The fix works by
having the client explicitly publish a unit name in the
relation data. That unit name is then used by the client
when looking up its certificates in the returned dict.

*1 https://github.com/juju-solutions/interface-tls-certificates/commit/231c86ce4086fa4d3d78414594336089b519f704