Add service_{project,domain}_id keys to Ident ctxt (#346)
* Add service_{project,domain}_id keys to Ident ctxt

Add service_{project,domain}_id keys to IdentityServiceContext. The
data is already present in the admin_domain_name and admin_tenant_id
keys but those key names are confusing.

Adding service_{project,domain}_id keys to the context is gated on
the use of keystone api version greater than 2. It is assumed that
if api 3 is being used then the service project and domain ids
are being passed down the relation which is true as of the 18.04
keystone charms (https://review.opendev.org/#/c/548264).

* Restore deepcopy