Add getter helpers to contrib ovs module (#353)
* Add getter helpers to contrib ovs module

This patch adds in some getters for listing bridges and ports on bridges
on the openvswitch.  This complements the add and delete methods and is
(primarily) to allow the neutron-gateway charm to remove bridges/ports
that are no longer in the config for that application.

Related-Bug: lp#1809190

* Update test_ovs.py for PEP8 compliance

* Filter out empty strings if ovs-vsctl returns no lines

If there are no bridges configured, then the command may return no lines
and thus would return an empty string.  Thus, filter them out, so that
an empty list is returned instead.