Update the logic for getting Juju version
Since ~ Juju 2.0, the JUJU_VERSION env var has provided the charm info
about what version of Juju the charm is running under.  The old hack may
not work going forward (particularly on Kubernetes charms).

One caveat of this change is that the env var doesn't include the series
and arch info, but I doubt anyone was actually depending on that.