fetch: Override apt execution environment
Override the execution environment for the apt_* family of
functions with a sanitized environment.

The need for this rises from consumers of the library running the
code from within a virtualenv.  In such a situation a deb
post-install script interacting with Python may interact with the
virtualenv instead of the system Python environment.

Clean up a pre-existing left behind unit test change after update
to _run_with_retries().

LP: #1840257