Adding duplicate check for ops.
This came about due to changes to ceph ops to allow creating multiple pools.
The thought of multiple calls with the same pool names came up. While this
prevents multiple ops from the same relation, it isn't going to prevent two
different applications from attempting to create multiple pools. The point
of that comment is to point out that deplication is still possible.