Add simple template facility to policy.d override
Adds a simple template facility so that a charm can choose whether to
perform substitutions on the documents loaded by the policy.d override
framework.  This is NOT Jinja templating, although there is nothing to
stop the charm from doing that.  It is simply a hook within the policy.d
loading mechanism so that a charm *can* perform templating in a
relatively simple way.

Also additional documentation about how to incorporate the module into a
charm.