Add policyd override helpers (#368)
These helpers implement the policy.d override functionality for
OpenStack charms.

* resolve circular import

* Fix os.isfile -> os.path.isfile

* Fix openstack_release handling

* Ensure path for policyd exists

* Another minor syntax error

* Further changes to get policyd working

* Add tests for the policyd overrides support code

* Add in the always blacklisted keys

The policy keys "admin_required" and "cloud_admin" are protected, in
that the default policy for a charm is typically rendered with these
keys.  This patch is to prevent breakage in an OpenStack juju deployed
system by overwriting these keys.

* Fix PEP8 errors added in last commit

* Add simple template facility to policy.d override

Adds a simple template facility so that a charm can choose whether to
perform substitutions on the documents loaded by the policy.d override
framework.  This is NOT Jinja templating, although there is nothing to
stop the charm from doing that.  It is simply a hook within the policy.d
loading mechanism so that a charm *can* perform templating in a
relatively simple way.

Also additional documentation about how to incorporate the module into a
charm.

* Change supported ubuntu release from bionic to xenial

It turns out that there is a requirement to support  policy
overrides at xenial/queens, and not just bionic/queens.

* Add in additional template extensions and fix bug in yaml_files

* Ensure that if config is False that policies are removed

This patch ensures that the policyd directory is definitely
cleared if the config value is False.

* Simplify the test for validity

It was pointed out that queens is only available on xenial onwards, so a
test being newer than xenial was redundant.  This patch removes the
redundant check and updates the test, which also has the benefit of
being simpler.