Fix policyd helper where when the config value is set to false
The issue was that the restart helper wasn't called when config was set
to false and previously there had been a working policy override.  If
the underlying service needs a restart (e.g. neutron-server) then the
policy override change isn't seen, and so the policy stays in force.
This change ensures the restart handler is called under all
circumstances.