Fix policyd on trusty (py34 issue)
os.scandir() doesn't exit on py34 (trusty).  Trusty isn't a target for
policyd, but due to another change for the cinder charm (only) the
policyd valid test was in the wrong place, and it exposed the scandir
issue on py34.  This change puts the test in the right place *and* fixes
the trusty issue in case policyd *is* backported to a trusty release.