Rework vault token management (#401)
The code in contrib.openstack.vaultlocker that handles token
is not tolerant of eventual consistency across the remote
units of vault (i.e. the provider). This change allows the
consumer to try all available tokens until one succeeds and
if none succeeds it falls back to using the cached sceret-id
if avaiable otherwise marks the interface as incomplete.

Related-Bug: #1849323