Ensure python3-hvac is installed for charms with encypt option
The referenced bug is essentially: make vault:secrets relation to vault
but keep the 'encrypt' option as False. In this case, the Context
handling code in charm-helpers is expecting python3-hvac to be
available, but it is only installed if the encrypt option is set to
True.  Hence the charm crashes.  This resolves that crash.

The VautlKVContext()() returns {} if the hvac module isn't isntalled,
which essentially deactivates it.

Closes-bug: #1862085