add helpers for log with levels
currently to write an error log with charmhelpers.hookenv:

    hookenv.log('this is an error!', level=hookenv.ERROR)

The `level` arg is a bit overwhelmed for user.
Thus, it ends up ignored normally, and defaults to INFO even it's an ERROR in a lot of charms.

Add simple helpers, so we can write a log with correct level without thinking.

Signed-off-by: Joe Guo <joe.guo@canonical.com>