ovs: Add ``add_bridge_bond`` function
The function allows the user to add bonded port in bridge from
interfaces and optionally attach data to the port and interfaces.

Depending on the configuration provided through the ``portdata``
and ``ifdatamap`` configuration options this function can be used
for adding bonds comprised of any interface type supported by
Open vSwitch.