ovs/ovsdb: Improve SimpleOVSDB API
The current approach is string heavy and could fit better with how
Python developers expect data structures to behave.

As there has been no release of charm-helpers since the initial
commit of the API it should be safe to change it at this point.

A future improvement could still be to add a third dimension to
the data structure to allow users to operate on specific key/value
pairs in a specifc row in a specific table.  At this point in time
the underlying command line argument that would be used does not
support outputting data in JSON format so for that reason I'm
postponing implementation of that facet.