Add hook context to unit test helper
Add support to tests.helper.FakeRelation for setting a implicit
relation context. Also add support for relation_id. This exposed
a mistake in two other unit tests, so fixes to apacheutils tests
included too.

set_relation_context can be used to simulate being in relation hook
context, eg omitting a relation id or unit implies the query is
against the relation id and unit in the current hook context. To set
the context:

    relation = FakeRelation(rel)
    relation.set_relation_context('mysql-svc2/0', 'shared-db:12')

To clear it:

    relation.clear_relation_context()