Move state check from keystone (#453)
* Add hook context to unit test helper

Add support to tests.helper.FakeRelation for setting a implicit
relation context. Also add support for relation_id. This exposed
a mistake in two other unit tests, so fixes to apacheutils tests
included too.

set_relation_context can be used to simulate being in relation hook
context, eg omitting a relation id or unit implies the query is
against the relation id and unit in the current hook context. To set
the context:

    relation = FakeRelation(rel)
    relation.set_relation_context('mysql-svc2/0', 'shared-db:12')

To clear it:

    relation.clear_relation_context()

* Improve docstring

* Port 3 state check helpers from keystone to ch

This change copies over container_scoped_relations, is_db_ready and
is_expected_scale from the keystone charm *1. The functions have
been refactered a little. The only change in functionality is that
is_db_ready no longer reports True in the absence of a database
relation.

*1 https://github.com/openstack/charm-keystone/blob/master/hooks/keystone_utils.py#L2576