non-active local unit state in application status
The current code causes any non-active state of the local unit and
its correspoding message to be lost when setting application work
load status. This change, as suggested by cory_fu, allows a
non-active state for this unit to be displayed as it now takes
precedence over a peer not being ready.