openstack: Move bridge port interface mapping to c-h
This will allow uniform handling of network configuration
across ``neutron-openvswitch``, ``neutron-gateway`` and
the OVN chassis charms.

Note that the handling of this detail in the charm is
pre-deprecated.

The long term goal is for network connectivity detail to be
modelled in the server provisioning layer (such as MAAS) which in
turn will provide a Netplan YAML description that will be used to
drive Open vSwitch.

Until we get to that reality the charm will need to configure
this detail based on application level configuration options.

Rename and extend the ``BridgeBondMap`` class to
``BridgePortInterfaceMap`` and use the same data structures
internally for regular bridge interface mapping and the DPDK OVS
bonding support. Note in light of the pre-deprecation mentioned
above we only expose OVS bond configuration for the DPDK use case.