ovs: Allow ``add_bridge_port`` to leave promisc untouched
At present the function will actively turn promisc mode on or off
depending on the ``promisc`` argument.

This does not play nice with operating on ports that do not exist
or are visible to userland.

Make the ``promisc`` argument optional and do not attempt to set
it if argument is not provided. The default of 'False' is
retained for compatibility.